+++ uctp-modules/uctp-business/src/main/java/com/newtouch/uctp/module/business/dal/dataobject/profit/MerchantProfitDO.java
@@ -31,9 +31,9 @@ public class MerchantProfitDO extends BaseDO {
     private String tradeTo; // 交易去向
     private String tradeToText; // 交易去向中文名称
     private String contractNo; // 交易合同号
-    private Integer profit; // 利润
-    private Integer profitBalance; // 利润余额
-    private Integer cashBack; // 待回填保证金额度
+    private Long profit; // 利润
+    private Long profitBalance; // 利润余额
+    private Long cashBack; // 待回填保证金额度
     private String bankNo; // 提现银行卡号
     private String bankName;// 提现银行名称
     private String presentState;// 提现状态
